[MOD] Modified LeaderBoard
Modified it so it has a reset button to reset the leaderboard. Also made a proper pathFile for the LeaederBoard.txt.
Everything works the game builds and runs properly, extra things that I could do was build more levels to the game but not planning on making the actual whole game just a demo.